Photos/Version-2/dev/requirements.txt
# Google Sheets Type Detection and Conversion Module Requirements

# Core Google API libraries
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0

# Authentication and HTTP handling
google-auth>=2.0.0
requests>=2.25.0

# Data processing (optional - for CSV operations)
pandas>=1.3.0

# Type hints support
typing-extensions>=4.0.0

# Development and testing (optional)
pytest>=6.0.0
pytest-mock>=3.0.0
